+++ plugins/wp-react-blocks-plugin/blocks/posts-filter/BlockEdit.jsx
@@ -57,6 +57,7 @@ class BlockEdit extends BlockEditWithFilters {
                 isYearFilter,
                 selectedYear,
                 defaultValues,
+                wordpressSource,
 
             }
         } = this.props;
@@ -76,13 +77,14 @@ class BlockEdit extends BlockEditWithFilters {
                         </PanelRow>
                     </PanelBody>
 
-                    <PanelBody title={__("Filter Configuration")}>
+                    {this.renderWordpressSource()}
 
+                    <PanelBody title={__("Filter Configuration")}>
                         <SelectControl
                             label={__('Filter Type')}
                             value={filterType}
                             onChange={(filterType) => {
-                                setAttributes({ filterType: filterType });
+                                setAttributes({ filterType: filterType, defaultValues: [] });
                             }}
                             options={[
                                 { label: "Single select", value: "single-select" },
@@ -97,7 +99,8 @@ class BlockEdit extends BlockEditWithFilters {
                             onChange={() => setAttributes({
                                 isCountryFilter: !isCountryFilter,
                                 isYearFilter: false,
-                                selectedYear: null
+                                selectedYear: null,
+                                defaultValues: []
                             })} />
 
                         <ToggleControl
@@ -107,7 +110,8 @@ class BlockEdit extends BlockEditWithFilters {
                             onChange={() => setAttributes({
                                 isYearFilter: !isYearFilter,
                                 isCountryFilter: false,
-                                selectedYear: null
+                                selectedYear: null,
+                                defaultValues: []
                             })} />
 
                     </PanelBody>