+++ plugins/wp-react-blocks-plugin/blocks/posts-pagination/BlockSave.jsx
@@ -4,7 +4,11 @@ const BlockSave = (props) => {
     const {
         attributes: {
             group,
-            numberOfItemsPerPage
+            numberOfItemsPerPage,
+            wordpressSource,
+            wordpressSourceType,
+            pageLabel,
+            ofLabel,
         }
     } = props;
 
@@ -18,6 +22,10 @@ const BlockSave = (props) => {
             data-component={"postsPagination"}
             data-group={group}
             data-number-of-items-per-page={numberOfItemsPerPage}
+            data-wordpress-source={wordpressSource}
+            data-wordpress-source-type={wordpressSourceType}
+            data-page-label={pageLabel}
+            data-of-label={ofLabel}
         ></div>
     );
 };