+++ plugins/wp-react-blocks-plugin/blocks/posts-pagination/index.js
@@ -18,6 +18,22 @@ registerBlockType(BLOCKS_NS + '/posts-pagination',
             numberOfItemsPerPage: {
                 type: 'number',
                 default: 10,
+            },
+            wordpressSource: {
+                type: 'string',
+                default: ''
+            },
+            wordpressSourceType: {
+                type: 'string',
+                default: 'internal'
+            },
+            pageLabel: {
+                type: 'string',
+                default: 'Page'
+            },
+            ofLabel: {
+                type: 'string',
+                default: 'of'
             }
         },
         edit: BlockEdit,