+++ wp-theme/page.php
@@ -1,5 +1,5 @@
 <?php
-$preview = $_GET['$preview'];
+$preview = sanitize_text_field($_GET['preview'] ?? '');
 global $post;
  session_start();
 $POST_URL=  str_replace(get_site_url(),"",get_permalink($post));
@@ -8,7 +8,7 @@ if ( str_starts_with($POST_URL,"/".wpm_get_language())){
 }
 if ( is_preview() ) {
 
-    wp_redirect(get_option( 'react_ui_url' )."/".$_SESSION["page_locale"]."/preview/".get_post_type($post)."/". $post->ID."?_wpnonce="."".wp_create_nonce( "wp_rest")."&preview="."".$_GET["preview"]);
+    wp_redirect(get_option( 'react_ui_url' )."/".$_SESSION["page_locale"]."/preview/".get_post_type($post)."/". $post->ID."?_wpnonce="."".wp_create_nonce( "wp_rest")."&preview=".$preview);
  }else{
     wp_redirect(get_option( 'react_ui_url' )."/".wpm_get_language().$POST_URL);
 }